requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
carlschader_ml_utils==0.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
fiblat==0.5.0
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
grpcio==1.62.2
huggingface-hub==0.22.2
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.3
lightning==2.2.2
lightning-utilities==0.11.2
llvmlite==0.42.0
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
protobuf==5.26.1
py-cpuinfo==9.0.0
pytest==8.1.1
pytest-benchmark==4.0.0
pytorch-lightning==2.2.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tokenizers==0.19.1
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.0
triton==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
